@@ -13,11 +13,11 @@ class Chevrons extends Pattern {
   final Color strokeColor;
 
   Chevrons(
-      {@required this.side,
-      @required this.nx,
-      @required this.ny,
-      @required this.fillColors,
-      @required this.strokeColor})
+      {required this.side,
+      required this.nx,
+      required this.ny,
+      required this.fillColors,
+      required this.strokeColor})
       : assert(fillColors.length == nx * ny);
 
   Chevrons.fromHash(String hash)
